+++ src/ophyd_async/tango/testing/_example_types.py
@@ -1,0 +1,7 @@
+from ophyd_async.core import StrictEnum
+
+
+class ExampleStrEnum(StrictEnum):
+    A = "AAA"
+    B = "BBB"
+    C = "CCC"
