+++ src/ophyd_async/tango/testing/_tango/__init__.py
@@ -1,0 +1,3 @@
+from ._one_of_everything import OneOfEverythingTangoDevice
+
+__all__ = ["OneOfEverythingTangoDevice"]
