+++ tests/system_tests_tango/test_tango_signals.py
@@ -30,8 +30,10 @@ from ophyd_async.tango.core import (
 )
 from ophyd_async.tango.testing import (
     ExampleStrEnum,
-    OneOfEverythingTangoDevice,
     TangoSubprocessDeviceServer,
+)
+from ophyd_async.tango.testing._tango import (  # noqa: PLC2701
+    OneOfEverythingTangoDevice,
 )
 from ophyd_async.testing import (
     MonitorQueue,
